Struct aws_sdk_appsync::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSDeepdishControlPlaneService.

This client allows ergonomic access to a AWSDeepdishControlPlaneService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_api_cache(&self) -> CreateApiCache<C, M, R>

Constructs a fluent builder for the CreateApiCache operation.

See CreateApiCache for more information about the operation and its arguments.

pub fn create_api_key(&self) -> CreateApiKey<C, M, R>

Constructs a fluent builder for the CreateApiKey operation.

See CreateApiKey for more information about the operation and its arguments.

pub fn create_data_source(&self) -> CreateDataSource<C, M, R>

Constructs a fluent builder for the CreateDataSource operation.

See CreateDataSource for more information about the operation and its arguments.

pub fn create_function(&self) -> CreateFunction<C, M, R>

Constructs a fluent builder for the CreateFunction operation.

See CreateFunction for more information about the operation and its arguments.

pub fn create_graphql_api(&self) -> CreateGraphqlApi<C, M, R>

Constructs a fluent builder for the CreateGraphqlApi operation.

See CreateGraphqlApi for more information about the operation and its arguments.

pub fn create_resolver(&self) -> CreateResolver<C, M, R>

Constructs a fluent builder for the CreateResolver operation.

See CreateResolver for more information about the operation and its arguments.

pub fn create_type(&self) -> CreateType<C, M, R>

Constructs a fluent builder for the CreateType operation.

See CreateType for more information about the operation and its arguments.

pub fn delete_api_cache(&self) -> DeleteApiCache<C, M, R>

Constructs a fluent builder for the DeleteApiCache operation.

See DeleteApiCache for more information about the operation and its arguments.

pub fn delete_api_key(&self) -> DeleteApiKey<C, M, R>

Constructs a fluent builder for the DeleteApiKey operation.

See DeleteApiKey for more information about the operation and its arguments.

pub fn delete_data_source(&self) -> DeleteDataSource<C, M, R>

Constructs a fluent builder for the DeleteDataSource operation.

See DeleteDataSource for more information about the operation and its arguments.

pub fn delete_function(&self) -> DeleteFunction<C, M, R>

Constructs a fluent builder for the DeleteFunction operation.

See DeleteFunction for more information about the operation and its arguments.

pub fn delete_graphql_api(&self) -> DeleteGraphqlApi<C, M, R>

Constructs a fluent builder for the DeleteGraphqlApi operation.

See DeleteGraphqlApi for more information about the operation and its arguments.

pub fn delete_resolver(&self) -> DeleteResolver<C, M, R>

Constructs a fluent builder for the DeleteResolver operation.

See DeleteResolver for more information about the operation and its arguments.

pub fn delete_type(&self) -> DeleteType<C, M, R>

Constructs a fluent builder for the DeleteType operation.

See DeleteType for more information about the operation and its arguments.

pub fn flush_api_cache(&self) -> FlushApiCache<C, M, R>

Constructs a fluent builder for the FlushApiCache operation.

See FlushApiCache for more information about the operation and its arguments.

pub fn get_api_cache(&self) -> GetApiCache<C, M, R>

Constructs a fluent builder for the GetApiCache operation.

See GetApiCache for more information about the operation and its arguments.

pub fn get_data_source(&self) -> GetDataSource<C, M, R>

Constructs a fluent builder for the GetDataSource operation.

See GetDataSource for more information about the operation and its arguments.

pub fn get_function(&self) -> GetFunction<C, M, R>

Constructs a fluent builder for the GetFunction operation.

See GetFunction for more information about the operation and its arguments.

pub fn get_graphql_api(&self) -> GetGraphqlApi<C, M, R>

Constructs a fluent builder for the GetGraphqlApi operation.

See GetGraphqlApi for more information about the operation and its arguments.

pub fn get_introspection_schema(&self) -> GetIntrospectionSchema<C, M, R>

Constructs a fluent builder for the GetIntrospectionSchema operation.

See GetIntrospectionSchema for more information about the operation and its arguments.

pub fn get_resolver(&self) -> GetResolver<C, M, R>

Constructs a fluent builder for the GetResolver operation.

See GetResolver for more information about the operation and its arguments.

pub fn get_schema_creation_status(&self) -> GetSchemaCreationStatus<C, M, R>

Constructs a fluent builder for the GetSchemaCreationStatus operation.

See GetSchemaCreationStatus for more information about the operation and its arguments.

pub fn get_type(&self) -> GetType<C, M, R>

Constructs a fluent builder for the GetType operation.

See GetType for more information about the operation and its arguments.

pub fn list_api_keys(&self) -> ListApiKeys<C, M, R>

Constructs a fluent builder for the ListApiKeys operation.

See ListApiKeys for more information about the operation and its arguments.

pub fn list_data_sources(&self) -> ListDataSources<C, M, R>

Constructs a fluent builder for the ListDataSources operation.

See ListDataSources for more information about the operation and its arguments.

pub fn list_functions(&self) -> ListFunctions<C, M, R>

Constructs a fluent builder for the ListFunctions operation.

See ListFunctions for more information about the operation and its arguments.

pub fn list_graphql_apis(&self) -> ListGraphqlApis<C, M, R>

Constructs a fluent builder for the ListGraphqlApis operation.

See ListGraphqlApis for more information about the operation and its arguments.

pub fn list_resolvers(&self) -> ListResolvers<C, M, R>

Constructs a fluent builder for the ListResolvers operation.

See ListResolvers for more information about the operation and its arguments.

pub fn list_resolvers_by_function(&self) -> ListResolversByFunction<C, M, R>

Constructs a fluent builder for the ListResolversByFunction operation.

See ListResolversByFunction for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_types(&self) -> ListTypes<C, M, R>

Constructs a fluent builder for the ListTypes operation.

See ListTypes for more information about the operation and its arguments.

pub fn start_schema_creation(&self) -> StartSchemaCreation<C, M, R>

Constructs a fluent builder for the StartSchemaCreation operation.

See StartSchemaCreation for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_api_cache(&self) -> UpdateApiCache<C, M, R>

Constructs a fluent builder for the UpdateApiCache operation.

See UpdateApiCache for more information about the operation and its arguments.

pub fn update_api_key(&self) -> UpdateApiKey<C, M, R>

Constructs a fluent builder for the UpdateApiKey operation.

See UpdateApiKey for more information about the operation and its arguments.

pub fn update_data_source(&self) -> UpdateDataSource<C, M, R>

Constructs a fluent builder for the UpdateDataSource operation.

See UpdateDataSource for more information about the operation and its arguments.

pub fn update_function(&self) -> UpdateFunction<C, M, R>

Constructs a fluent builder for the UpdateFunction operation.

See UpdateFunction for more information about the operation and its arguments.

pub fn update_graphql_api(&self) -> UpdateGraphqlApi<C, M, R>

Constructs a fluent builder for the UpdateGraphqlApi operation.

See UpdateGraphqlApi for more information about the operation and its arguments.

pub fn update_resolver(&self) -> UpdateResolver<C, M, R>

Constructs a fluent builder for the UpdateResolver operation.

See UpdateResolver for more information about the operation and its arguments.

pub fn update_type(&self) -> UpdateType<C, M, R>

Constructs a fluent builder for the UpdateType operation.

See UpdateType for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.