Module aws_sdk_appsync::error

Modules

access_denied_exception

See AccessDeniedException

api_key_limit_exceeded_exception

See ApiKeyLimitExceededException

api_key_validity_out_of_bounds_exception

See ApiKeyValidityOutOfBoundsException

api_limit_exceeded_exception

See ApiLimitExceededException

bad_request_exception

See BadRequestException

concurrent_modification_exception

See ConcurrentModificationException

graph_ql_schema_exception

See GraphQlSchemaException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

unauthorized_exception

See UnauthorizedException

Structs

AccessDeniedException

You do not have access to perform this operation on this resource.

ApiKeyLimitExceededException

The API key exceeded a limit. Try your request again.

ApiKeyValidityOutOfBoundsException

The API key expiration must be set to a value between 1 and 365 days from creation (for

ApiLimitExceededException

The GraphQL API exceeded a limit. Try your request again.

BadRequestException

The request is not well formed. For example, a value is invalid or a required field is

ConcurrentModificationException

Another modification is in progress at this time and it must complete before you can

CreateApiCacheError

CreateApiKeyError

CreateDataSourceError

CreateFunctionError

CreateGraphqlApiError

CreateResolverError

CreateTypeError

DeleteApiCacheError

DeleteApiKeyError

DeleteDataSourceError

DeleteFunctionError

DeleteGraphqlApiError

DeleteResolverError

DeleteTypeError

FlushApiCacheError

GetApiCacheError

GetDataSourceError

GetFunctionError

GetGraphqlApiError

GetIntrospectionSchemaError

GetResolverError

GetSchemaCreationStatusError

GetTypeError

GraphQlSchemaException

The GraphQL schema is not valid.

InternalFailureException

An internal AppSync error occurred. Try your request again.

LimitExceededException

The request exceeded a limit. Try your request again.

ListApiKeysError

ListDataSourcesError

ListFunctionsError

ListGraphqlApisError

ListResolversByFunctionError

ListResolversError

ListTagsForResourceError

ListTypesError

NotFoundException

The resource specified in the request was not found. Check the resource, and then try

StartSchemaCreationError

TagResourceError

UnauthorizedException

You are not authorized to perform this operation.

UntagResourceError

UpdateApiCacheError

UpdateApiKeyError

UpdateDataSourceError

UpdateFunctionError

UpdateGraphqlApiError

UpdateResolverError

UpdateTypeError

Enums

CreateApiCacheErrorKind

CreateApiKeyErrorKind

CreateDataSourceErrorKind

CreateFunctionErrorKind

CreateGraphqlApiErrorKind

CreateResolverErrorKind

CreateTypeErrorKind

DeleteApiCacheErrorKind

DeleteApiKeyErrorKind

DeleteDataSourceErrorKind

DeleteFunctionErrorKind

DeleteGraphqlApiErrorKind

DeleteResolverErrorKind

DeleteTypeErrorKind

FlushApiCacheErrorKind

GetApiCacheErrorKind

GetDataSourceErrorKind

GetFunctionErrorKind

GetGraphqlApiErrorKind

GetIntrospectionSchemaErrorKind

GetResolverErrorKind

GetSchemaCreationStatusErrorKind

GetTypeErrorKind

ListApiKeysErrorKind

ListDataSourcesErrorKind

ListFunctionsErrorKind

ListGraphqlApisErrorKind

ListResolversByFunctionErrorKind

ListResolversErrorKind

ListTagsForResourceErrorKind

ListTypesErrorKind

StartSchemaCreationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApiCacheErrorKind

UpdateApiKeyErrorKind

UpdateDataSourceErrorKind

UpdateFunctionErrorKind

UpdateGraphqlApiErrorKind

UpdateResolverErrorKind

UpdateTypeErrorKind