Module aws_sdk_appsync::model

Modules

additional_authentication_provider

See AdditionalAuthenticationProvider

api_cache

See ApiCache

api_key

See ApiKey

authorization_config

See AuthorizationConfig

aws_iam_config

See AwsIamConfig

caching_config

See CachingConfig

cognito_user_pool_config

See CognitoUserPoolConfig

data_source

See DataSource

delta_sync_config

See DeltaSyncConfig

dynamodb_data_source_config

See DynamodbDataSourceConfig

elasticsearch_data_source_config

See ElasticsearchDataSourceConfig

function_configuration

See FunctionConfiguration

graphql_api

See GraphqlApi

http_data_source_config

See HttpDataSourceConfig

lambda_authorizer_config

See LambdaAuthorizerConfig

lambda_conflict_handler_config

See LambdaConflictHandlerConfig

lambda_data_source_config

See LambdaDataSourceConfig

log_config

See LogConfig

open_id_connect_config

See OpenIdConnectConfig

pipeline_config

See PipelineConfig

rds_http_endpoint_config

See RdsHttpEndpointConfig

relational_database_data_source_config

See RelationalDatabaseDataSourceConfig

resolver

See Resolver

sync_config

See SyncConfig

type

See Type

user_pool_config

See UserPoolConfig

Structs

AdditionalAuthenticationProvider

Describes an additional authentication provider.

ApiCache

The ApiCache object.

ApiKey

Describes an API key.

AuthorizationConfig

The authorization config in case the HTTP endpoint requires authorization.

AwsIamConfig

The Identity and Access Management configuration.

CachingConfig

The caching configuration for a resolver that has caching enabled.

CognitoUserPoolConfig

Describes an Amazon Cognito user pool configuration.

DataSource

Describes a data source.

DeltaSyncConfig

Describes a Delta Sync configuration.

DynamodbDataSourceConfig

Describes an Amazon DynamoDB data source configuration.

ElasticsearchDataSourceConfig

Describes an Elasticsearch data source configuration.

FunctionConfiguration

A function is a reusable entity. Multiple functions can be used to compose the resolver

GraphqlApi

Describes a GraphQL API.

HttpDataSourceConfig

Describes an HTTP data source configuration.

LambdaAuthorizerConfig

A LambdaAuthorizerConfig holds configuration on how to authorize AppSync API access when using

LambdaConflictHandlerConfig

The LambdaConflictHandlerConfig object when configuring LAMBDA as the

LambdaDataSourceConfig

Describes an Amazon Web Services Lambda data source configuration.

LogConfig

The CloudWatch Logs configuration.

OpenIdConnectConfig

Describes an OpenID Connect configuration.

PipelineConfig

The pipeline configuration for a resolver of kind PIPELINE.

RdsHttpEndpointConfig

The Amazon RDS HTTP endpoint configuration.

RelationalDatabaseDataSourceConfig

Describes a relational database data source configuration.

Resolver

Describes a resolver.

SyncConfig

Describes a Sync configuration for a resolver.

Type

Describes a type.

UserPoolConfig

Describes an Amazon Cognito user pool configuration.

Enums

ApiCacheStatus

ApiCacheType

ApiCachingBehavior

AuthenticationType

AuthorizationType

ConflictDetectionType

ConflictHandlerType

DataSourceType

DefaultAction

FieldLogLevel

OutputType

RelationalDatabaseSourceType

ResolverKind

SchemaStatus

TypeDefinitionFormat