Module aws_sdk_appsync::operation

Structs

CreateApiCache

Creates a cache for the GraphQL API.

CreateApiKey

Creates a unique key that you can distribute to clients who are executing your

CreateDataSource

Creates a DataSource object.

CreateFunction

Creates a Function object.

CreateGraphqlApi

Creates a GraphqlApi object.

CreateResolver

Creates a Resolver object.

CreateType

Creates a Type object.

DeleteApiCache

Deletes an ApiCache object.

DeleteApiKey

Deletes an API key.

DeleteDataSource

Deletes a DataSource object.

DeleteFunction

Deletes a Function.

DeleteGraphqlApi

Deletes a GraphqlApi object.

DeleteResolver

Deletes a Resolver object.

DeleteType

Deletes a Type object.

FlushApiCache

Flushes an ApiCache object.

GetApiCache

Retrieves an ApiCache object.

GetDataSource

Retrieves a DataSource object.

GetFunction

Get a Function.

GetGraphqlApi

Retrieves a GraphqlApi object.

GetIntrospectionSchema

Retrieves the introspection schema for a GraphQL API.

GetResolver

Retrieves a Resolver object.

GetSchemaCreationStatus

Retrieves the current status of a schema creation operation.

GetType

Retrieves a Type object.

ListApiKeys

Lists the API keys for a given API.

ListDataSources

Lists the data sources for a given API.

ListFunctions

List multiple functions.

ListGraphqlApis

Lists your GraphQL APIs.

ListResolvers

Lists the resolvers for a given API and type.

ListResolversByFunction

List the resolvers that are associated with a specific function.

ListTagsForResource

Lists the tags for a resource.

ListTypes

Lists the types for a given API.

StartSchemaCreation

Adds a new schema to your GraphQL API.

TagResource

Tags a resource with user-supplied tags.

UntagResource

Untags a resource.

UpdateApiCache

Updates the cache for the GraphQL API.

UpdateApiKey

Updates an API key. The key can be updated while it is not deleted.

UpdateDataSource

Updates a DataSource object.

UpdateFunction

Updates a Function object.

UpdateGraphqlApi

Updates a GraphqlApi object.

UpdateResolver

Updates a Resolver object.

UpdateType

Updates a Type object.